requirements.txt
dvc[s3]==3.55.2
scikit-learn==1.5.2
pydantic==2.9.1
pytorch-lightning==2.4.0
wandb==0.18.0
omegaconf==2.3.0

bitsandbytes==0.43.3
datasets==3.0.1
accelerate==0.34.2
transformers==4.45.1

tensorboard==2.17.0
pandas==2.2.2
polars==1.7.1
seaborn>=0.13.2
tqdm>=4.66.4
typer>=0.12.5
tabulate>=0.9.0
umap-learn==0.5.5
hydra-core==1.3.2
loguru==0.7.2
python-dotenv==1.0.1

jupyterlab>=4.2.5
ipywidgets>=8.1.5

pre-commit>=3.8.0
mypy>=1.11.2
pytest==8.3.3
